requirements.txt
DateTime==4.3
numpy==1.20.3
pandas==1.2.4
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
zope.interface==5.4.0
